🚚 Rename : type 패키지로 status enum 이동
