+++ src/main/java/com/gogym/member/dto/SignUpRequest.java
@@ -1,6 +1,7 @@
 package com.gogym.member.dto;
 
 import com.gogym.member.entity.Role;
+import com.gogym.member.type.MemberStatus;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -44,6 +45,6 @@ public class SignUpRequest {
     return Member.builder().name(this.name).email(this.email).nickname(this.nickname)
         .phone(this.phone).password(encodedPassword).role(Role.USER)
         .profileImageUrl(this.profileImageUrl).regionId1(this.regionId1).regionId2(this.regionId2)
-        .build();
+        .memberStatus(MemberStatus.ACTIVE).build();
   }
 }
