+++ src/main/java/com/gogym/member/service/AuthService.java
@@ -6,10 +6,10 @@ import com.gogym.member.dto.ResetPasswordRequest;
 import com.gogym.member.dto.SignInRequest;
 import com.gogym.member.dto.SignUpRequest;
 import com.gogym.member.entity.Member;
-import com.gogym.member.entity.MemberStatus;
 import com.gogym.member.jwt.JwtTokenProvider;
 import com.gogym.member.repository.BanNicknameRepository;
 import com.gogym.member.repository.MemberRepository;
+import com.gogym.member.type.MemberStatus;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -40,8 +40,6 @@ public class AuthService {
 
     // Dto에서 Entity 변환
     Member member = request.toEntity(passwordEncoder.encode(request.getPassword()));
-
-    member.setMemberStatus(MemberStatus.ACTIVE);
 
     // 회원 데이터 저장
     memberRepository.save(member);
