+++ src/test/java/com/gogym/member/service/MemberServiceTest.java
@@ -7,9 +7,9 @@ import com.gogym.exception.CustomException;
 import com.gogym.member.dto.MemberProfileResponse;
 import com.gogym.member.dto.UpdateMemberRequest;
 import com.gogym.member.entity.Member;
-import com.gogym.member.entity.MemberStatus;
 import com.gogym.member.repository.BanNicknameRepository;
 import com.gogym.member.repository.MemberRepository;
+import com.gogym.member.type.MemberStatus;
 import com.gogym.gympay.entity.GymPay;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeEach;
